experiments/requirements.txt
scipy==1.5.2
wand==0.5.8
scikit-image==0.16.2
opencv-python==4.1.2.30
simplejson==3.17.0
pillow==6.2.2
numpy==1.19.2
hydra-core==1.0.6
submitit==1.2.0
hydra-submitit-launcher==1.1.0
